fixed large file bugs

+++ server.py
@@ -2,11 +2,13 @@
 # coding=utf-8
 
 import os
+import shutil
 from flask import Flask, request, Response,redirect, render_template as rt
+from itsdangerous import exc
 app = Flask(__name__)
 
 
-saveDir = '/home/'
+saveDir = './upload/'
 # @simple_page.route('/', defaults={'page': 'index'})
 # @simple_page.route('/<page>')
 @app.route('/',defaults={'fileName': None}, methods=['GET'])
@@ -17,7 +19,8 @@ def index(fileName):
     url = request.args.get('url', None)
     print("url is ",url)
     print("saveDir is ",saveDir)
-    if url != None:
+    print("changeDir is ",changeDir)
+    if url != None and changeDir:
         if os.path.isdir(url):
             saveDir = str(url)
             print("saveDir is ",saveDir)
@@ -28,7 +31,6 @@ def index(fileName):
 # def file_download(fileName):
     print("fileName is ",fileName)
     # print(os.path.isdir(fileName))
-
     if os.path.isdir(saveDir + fileName):
         if fileName[-1] != '/':
             return redirect('/'+fileName + '/')
@@ -95,15 +97,21 @@ def index(fileName):
 @app.route('/',defaults={'fileName': None}, methods=['POST'])
 @app.route('/<path:fileName>', methods=['POST'])
 def upload_part(fileName):  # 接收前端上传的一个分片
-    saveDirN = saveDir
-    print("fileName is ",fileName)
+    # print("request.form is \n",request.form)
+    nowDirN = saveDir
+    # print("fileName is ",fileName)
     if fileName != None:
-        saveDirN = saveDir + '/' + fileName + '/'
+        nowDirN = saveDir + '/' + fileName + '/'
     saveName = request.form.get('name')  # 获取文件的唯一标识符
     chunks = request.form.get('chunks')  # 获取该分片在所有分片中的序号
     chunk = request.form.get('chunk')  # 获取该分片在所有分片中的序号
+    try:os.mkdir((nowDirN+saveName+'.Save'),755)
+    except Exception as e:print(e)
+    saveDirN = nowDirN + (saveName+'.Save' + '/') 
     if chunks != None:
-        saveName = '%s_%s' % (saveName, chunk)  # 构造该分片的唯一标识符
+        saveName = '%s_%s' % (saveName, chunk)  # 构造该分片的唯一标识符  
+        os.rmdir(saveDirN)
+        saveDirN = nowDirN
         # saveName = request.form.get('name')  # 获取文件的唯一标识符
     upload_file = request.files['file']
     upload_file.save(saveDirN + '%s' % saveName)  # 保存分片到本地
@@ -118,21 +126,52 @@ def upload_part(fileName):  # 接收前端上传的一个分片
     target_fileName = request.form.get('name')  # 获取上传文件的文件名
     # task = request.args.get('task_id')  # 获取文件的唯一标识符
     # chunk = 0  # 分片序号
-    chunk = int(chunk)  # 分片序号
-    with open(saveDirN + '%s' % target_fileName, 'ab') as target_file:  # 创建新文件
-        try:
-            fileName = saveDirN + '%s_%s' % (target_fileName, chunk)
-            source_file = open(fileName, 'rb')  # 按序打开每个分片
-            target_file.write(source_file.read())  # 读取分片内容写入新文件
-            source_file.close()
-            os.remove(fileName)  # 删除该分片，节约空间
-        except Exception as e:
-            print("error is ",e)
-    
-            
-    print("upload all chunks saveas ",target_fileName)
-    return str(0)
+    # print((os.listdir(saveDirN)))
+    # print(len(os.listdir(saveDirN)))
+    # print(chunks)
+    # print(len(os.listdir(saveDirN)) == chunks)
+    if len(os.listdir(saveDirN)) == int(chunks):
+        print("All file chunks got ",saveDirN + '%s' % target_fileName)
+        chunk = 0
+        with open(saveDirN + '%s' % target_fileName, 'wb') as target_file:  # 创建新文件
+            while True:
+                try:
+                    fileName = saveDirN + '%s_%s' % (target_fileName, chunk)
+                    source_file = open(fileName, 'rb')  # 按序打开每个分片
+                    target_file.write(source_file.read())  # 读取分片内容写入新文件
+                    source_file.close()
+                    os.remove(fileName)  # 删除该分片，节约空间
+                except Exception as e:
+                    print("error is ",e)
+                    break
+                chunk += 1
+        shutil.move(saveDirN+target_fileName,nowDirN+target_fileName)
+        os.rmdir(saveDirN)
+        print("upload all chunks save as ",target_fileName)
+        return str(1)
+    else:
+        return str(0)
     # return rt('./index.html')
 
+import sys
+
+changeDir = False
+def mainApp():
+    print("Usage py server.py saveDir ifCanChangeDir port")
+    host = '0.0.0.0'
+    port = '8765'
+    debugT = False
+    print(sys.argv)
+    try:
+        global changeDir 
+        global saveDir
+        saveDir = sys.argv[1]
+        changeDir = sys.argv[2]
+        port = sys.argv[3]
+    except Exception as e:
+        pass
+    print('saveDir is :\n',saveDir,'\nIf or not can change dir:\n',changeDir,'\n')
+    app.run(host=host,port=port,debug = debugT)
+
 if __name__ == '__main__':
-    app.run(host='0.0.0.0',port=8765,debug = False)
+    mainApp()
